requirements.txt
fastapi
uvicorn
python-multipart
pillow
openai
google-genai
python-dotenv
